Refactor xtask crate
