+++ xtask/src/dist.rs
@@ -1,15 +1,10 @@
-use std::{
-    env,
-    fs::{self, File},
-    io::BufReader,
-    path::Path,
-    process::{Command, Stdio},
-};
+use std::{fs, path::Path};
 
-use cargo_metadata::{camino::Utf8PathBuf, Message, Metadata, MetadataCommand, Package};
+use cargo_metadata::{camino::Utf8PathBuf, Metadata, MetadataCommand, Package};
 use clap::Parser;
-use color_eyre::eyre::{ensure, Result};
-use flate2::{write::GzEncoder, Compression};
+use color_eyre::eyre::Result;
+
+use crate::util;
 
 #[derive(Debug, Parser)]
 pub struct Args {
@@ -38,86 +33,18 @@ pub fn run(args: &Args) -> Result<()> {
     let metadata = MetadataCommand::new().exec()?;
     let root_package = &metadata.root_package().unwrap();
 
-    let mut artifacts_path = vec![readme(&metadata)?];
-    for artifact in license(&metadata)? {
+    let mut artifacts_path = vec![util::collect_readme(root_package)?];
+    for artifact in util::collect_licenses(root_package)? {
         artifacts_path.push(artifact);
     }
-    for artifact in build(&metadata, root_package, use_cross, target)? {
+    for artifact in util::cargo_build_release_exe(&metadata, root_package, use_cross, target)? {
         artifacts_path.push(artifact);
     }
-    create_archive(&metadata, root_package, target, &artifacts_path)?;
+    let archive_path = create_archive(&metadata, root_package, target, &artifacts_path)?;
+
+    tracing::info!(%archive_path, "dist completed successfully");
 
     Ok(())
-}
-
-fn readme(metadata: &Metadata) -> Result<Utf8PathBuf> {
-    let path = metadata.workspace_root.join("README.md");
-    ensure!(path.is_file(), "README.md is not a file: {path}");
-    Ok(path)
-}
-
-fn license(metadata: &Metadata) -> Result<Vec<Utf8PathBuf>> {
-    let mut artifacts = vec![];
-    for ent in fs::read_dir(&metadata.workspace_root)? {
-        let ent = ent?;
-        let path = ent.path();
-        let file_name = path.file_name().unwrap().to_str().unwrap();
-        if file_name.starts_with("LICENSE-")
-            || file_name.starts_with("LICENSE.")
-            || file_name == "LICENSE"
-        {
-            ensure!(path.is_file(), "LICENSE is not a file: {}", path.display());
-            artifacts.push(path.try_into()?);
-        }
-    }
-    Ok(artifacts)
-}
-
-fn build(
-    metadata: &Metadata,
-    package: &Package,
-    use_cross: bool,
-    target: Option<&str>,
-) -> Result<Vec<Utf8PathBuf>> {
-    let cmd_name = if use_cross { "cross" } else { "cargo" };
-    let mut args = vec![
-        "build",
-        "--release",
-        "--package",
-        package.name.as_str(),
-        "--message-format=json-render-diagnostics",
-    ];
-
-    if let Some(target) = target {
-        args.extend_from_slice(&["--target", target]);
-    }
-
-    let mut cmd = Command::new(cmd_name);
-    cmd.args(&args);
-
-    tracing::info!("Running {} {}", cmd_name, args.join(" "));
-    let mut cmd = cmd.stdout(Stdio::piped()).spawn()?;
-
-    let mut artifacts = vec![];
-    let reader = BufReader::new(cmd.stdout.take().unwrap());
-    for message in Message::parse_stream(reader) {
-        if let Message::CompilerArtifact(msg) = message? {
-            if let Some(exe) = &msg.executable {
-                let exe = if use_cross {
-                    metadata
-                        .target_directory
-                        .join(exe.strip_prefix("/target").unwrap())
-                } else {
-                    exe.clone()
-                };
-                tracing::info!("Found artifact: {}", exe);
-                ensure!(exe.is_file(), "Artifact is not a file: {exe}");
-                artifacts.push(exe);
-            }
-        }
-    }
-
-    Ok(artifacts)
 }
 
 fn create_archive(
@@ -125,7 +52,7 @@ fn create_archive(
     package: &Package,
     target: Option<&str>,
     artifacts: &[impl AsRef<Path>],
-) -> Result<()> {
+) -> Result<Utf8PathBuf> {
     let dist_dir = metadata.target_directory.join("dist");
     fs::create_dir_all(&dist_dir)?;
 
@@ -136,20 +63,7 @@ fn create_archive(
 
     tracing::info!("Creating archive: {archive_path}");
 
-    let archive = File::create(&archive_path)?;
-    let enc = GzEncoder::new(archive, Compression::default());
-    let mut tar = tar::Builder::new(enc);
+    util::compress_files_tar_gz(&archive_path, artifacts)?;
 
-    for src_path in artifacts {
-        let src_path = src_path.as_ref();
-        let artifact_name = src_path.file_name().unwrap().to_str().unwrap();
-        tracing::info!(
-            "Adding artifact: {} as {}",
-            src_path.display(),
-            artifact_name
-        );
-        tar.append_path_with_name(src_path, artifact_name)?;
-    }
-
-    Ok(())
+    Ok(archive_path)
 }
